Changed psmod1 module project layout
Moving from: "The Play application layout"
to: "Default layout used by SBT and Maven"

Achieved adding .disablePlugins(PlayLayoutPlugin) to build.sbt
and moving all code and resources to their standard locations.

The goal is to lower barrier of entry to project by keeping
anything which is not the focus of the project as standard as
possible.